Rename TransactionList to SortedTransactionList. Only sort transaction once, and do not modify class internals in the getter method.
